Fixed a few errors in board piece matching points

+++ Source/F3MashUp/BoardPieceHolderCPP.h
@@ -63,7 +63,7 @@ public:
 	float RandomChangeFrequency;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	float PointsPerBoardPieceMatched;
+	int PointsPerBoardPieceMatched;
 
 	UFUNCTION(Server, Reliable)
 	void ServerRandomSwap();
